Make all formatting consistent
