+++ src/components/main/Main.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import {fetchScopedUsers} from './../../utils/fetchScopedUsers';
+import { fetchScopedUsers } from './../../utils/fetchScopedUsers';
 import PropTypes from 'prop-types';
 import firebase from './../../firebase';
-import ContactCardsContainer from '../../containers/ContactCardsContainer';
+import ContactCards from '../../components/contactCards/ContactCards';
 
 class Main extends React.Component {
   constructor() {
@@ -20,7 +20,7 @@ class Main extends React.Component {
 
     return loadedContacts.map((contact, index) => {
       return (
-        <ContactCardsContainer
+        <ContactCards
           key={index}
           name={contact.contactName}
         />
